+++ CHANGELOG.rst
@@ -4,6 +4,15 @@ Changelog
 Current release candidate
 -------------------------
 
+0.3.4 (2020-01-31)
+------------------
+
+* Fixed bug with deleting models from tasks
+* Support working with model meta without requiring installation of all model dependencies
+* Added region argument for s3 repository
+* Support for delete_model in Ebonite client
+* Support for force flag in delete_model which deletes model even if artifacts could not be deleted
+
 0.3.3 (2020-01-10)
 ------------------